requirements.txt
streamlit>=1.35.0
groq>=0.9.0
google-generativeai>=0.7.0
python-dotenv>=1.0.0
Pillow>=10.0.0
httpx>=0.27.0
requests>=2.31.0
sqlalchemy>=2.0.0
pydantic>=2.5.0
fastapi>=0.110.0
uvicorn>=0.27.0
langgraph>=0.2.0
chromadb>=0.4.0
reportlab>=4.0.0